+++ samples/unity-of-bugs/Assets/Editor/Builder.cs
@@ -8,10 +8,13 @@ using UnityEngine;
 
 public class Builder
 {
-    public static void BuildIl2CPPPlayer(BuildTarget buildTarget)
+    public static void BuildIl2CPPPlayer(BuildTarget target, BuildTargetGroup group)
     {
         var args = ParseCommandLineArguments();
         ValidateArguments(args);
+
+        // Make sure the configuration is right.
+        PlayerSettings.SetScriptingBackend(group, ScriptingImplementation.IL2CPP);
 
         if (args.ContainsKey("sentryOptions.configure"))
         {
@@ -20,9 +23,10 @@ public class Builder
 
         var buildPlayerOptions = new BuildPlayerOptions
         {
-            scenes = new[] {"Assets/Scenes/1_Bugfarm.unity"},
+            scenes = new[] { "Assets/Scenes/1_Bugfarm.unity" },
             locationPathName = args["buildPath"],
-            target = buildTarget,
+            target = target,
+            targetGroup = group,
             options = BuildOptions.StrictMode,
         };
 
@@ -58,10 +62,11 @@ public class Builder
         }
     }
 
-    public static void BuildWindowsIl2CPPPlayer() => BuildIl2CPPPlayer(BuildTarget.StandaloneWindows64);
-    public static void BuildMacIl2CPPPlayer() => BuildIl2CPPPlayer(BuildTarget.StandaloneOSX);
-    public static void BuildAndroidIl2CPPPlayer() => BuildIl2CPPPlayer(BuildTarget.Android);
-    public static void BuildIOSPlayer() => BuildIl2CPPPlayer(BuildTarget.iOS);
+    public static void BuildWindowsIl2CPPPlayer() => BuildIl2CPPPlayer(BuildTarget.StandaloneWindows64, BuildTargetGroup.Standalone);
+    public static void BuildMacIl2CPPPlayer() => BuildIl2CPPPlayer(BuildTarget.StandaloneOSX, BuildTargetGroup.Standalone);
+    public static void BuildLinuxIl2CPPPlayer() => BuildIl2CPPPlayer(BuildTarget.StandaloneLinux64, BuildTargetGroup.Standalone);
+    public static void BuildAndroidIl2CPPPlayer() => BuildIl2CPPPlayer(BuildTarget.Android, BuildTargetGroup.Android);
+    public static void BuildIOSPlayer() => BuildIl2CPPPlayer(BuildTarget.iOS, BuildTargetGroup.iOS);
 
     private static void SetupSentryOptions(Dictionary<string, string> args)
     {
@@ -79,7 +84,7 @@ public class Builder
         {
             foreach (var asm in AppDomain.CurrentDomain.GetAssemblies().Where(name => name.FullName.Contains("Sentry")))
             {
-                foreach(var type in asm.GetTypes())
+                foreach (var type in asm.GetTypes())
                 {
                     Debug.LogFormat(LogType.Log, LogOption.NoStacktrace, null, "SetupSentryOptions: Asm {0} Type {1}", asm.FullName, type.FullName);
                 }