+++ CHANGELOG.md
@@ -7,6 +7,12 @@
 - Samples include programmatic options configuration snippet ([#568](https://github.com/getsentry/sentry-unity/pull/568))
 - Support for programmatic options configuration ([#564](https://github.com/getsentry/sentry-unity/pull/564))
 
+### Fixes
+
+- Bump Sentry .NET SDK 3.14.1 ([#573](https://github.com/getsentry/sentry-unity/pull/573))
+  - [changelog](https://github.com/getsentry/sentry-dotnet/blob/3.14.1/CHANGELOG.md)
+  - [diff](https://github.com/getsentry/sentry-dotnet/compare/3.14.0...3.14.1)
+
 ## 0.10.0
 
 ### Features